backend/requirements.txt
altgraph                  0.17.5
annotated-types           0.7.0
anyio                     4.10.0
APScheduler               3.11.1
Authlib                   1.6.5
bcrypt                    4.0.0
bottle                    0.13.4
certifi                   2025.10.5
cffi                      2.0.0
charset-normalizer        3.4.4
click                     8.2.1
clr_loader                0.2.8
colorama                  0.4.6
cryptography              46.0.2
DateTime                  5.5
deprecation               2.1.0
dnspython                 2.8.0
email-validator           2.3.0
fastapi                   0.116.1
h11                       0.16.0
h2                        4.3.0
hpack                     4.1.0
httpcore                  1.0.9
httptools                 0.6.4
httpx                     0.28.1
hyperframe                6.1.0
idna                      3.10
itsdangerous              2.2.0
mailchimp-transactional   1.0.56
mailersend                2.0.0
packaging                 25.0
passlib                   1.7.4
pefile                    2024.8.26
pip                       24.0
postgrest                 2.21.1
proxy_tools               0.1.0
pycparser                 2.23
pydantic                  2.11.7
pydantic_core             2.33.2
pyinstaller               6.17.0
pyinstaller-hooks-contrib 2025.10
PyJWT                     2.10.1
pypaystack2               3.0.1
python-dateutil           2.9.0.post0
python-dotenv             1.1.1
python-multipart          0.0.20
pythonnet                 3.0.5
pytz                      2025.2
pywebview                 6.1
pywin32-ctypes            0.2.3
PyYAML                    6.0.2
realtime                  2.21.1
requests                  2.32.5
resend                    2.19.0
setuptools                65.5.0
six                       1.17.0
sniffio                   1.3.1
starlette                 0.47.2
storage3                  2.21.1
StrEnum                   0.4.15
supabase                  2.21.1
supabase-auth             2.21.1
supabase-functions        2.21.1
typing_extensions         4.14.1
typing-inspection         0.4.1
tzdata                    2025.2
tzlocal                   5.3.1
urllib3                   2.5.0
uvicorn                   0.35.0
watchfiles                1.1.0
websockets                15.0.1
zope.interface            8.0.1